+++ tools/benchmark/src/RunnerUtilities.ts
@@ -3,51 +3,13 @@
  * Licensed under the MIT License.
  */
 
-import type {
-	BenchmarkResult,
-	BenchmarkError,
-	BenchmarkData,
-	CustomData,
-	Stats,
-} from "./ResultTypes";
+import type { Stats } from "./ResultTypes";
 import { assert } from "./assert.js";
-import { timer } from "./timer";
 
 /**
  * This file contains generic utilities of use to a mocha reporter, especially for convenient formatting of textual
  * output to the command line.
  */
-
-/**
- * Wraps a function that returns CustomData, measuring its execution time
- * and capturing either its result or exception.
- * Returns a callback suitable for passing to emitResultsMocha.
- * This is a generic utility that is neither mocha-specific nor time benchmark-specific.
- */
-export function captureResults(
-	f: () => CustomData | Promise<CustomData>,
-): () => Promise<{ result: BenchmarkResult; exception?: Error }> {
-	return async () => {
-		const startTime = timer.now();
-
-		let customData: CustomData;
-		try {
-			customData = await f();
-		} catch (error) {
-			const benchmarkError: BenchmarkError = { error: (error as Error).message };
-			return { result: benchmarkError, exception: error as Error };
-		}
-
-		const elapsedSeconds = timer.toSeconds(startTime, timer.now());
-
-		const result: BenchmarkData = {
-			elapsedSeconds,
-			customData,
-		};
-
-		return { result };
-	};
-}
 
 /**
  * Creates and returns a padding string consisting of `num` copies of `chr`
