+++ tools/benchmark/src/mocha/customOutputRunner.ts
@@ -7,7 +7,8 @@ import type { Test } from "mocha";
 
 import type { BenchmarkDescription, MochaExclusiveOptions, Titled } from "../Configuration";
 import type { CustomData } from "../ResultTypes";
-import { prettyNumber, captureResults } from "../RunnerUtilities";
+import { captureResults } from "../ResultUtilities";
+import { prettyNumber } from "../RunnerUtilities";
 
 import { emitResultsMocha } from "./runnerUtilities";
